fix: address FUT-80 review hardening feedback
