+++ crates/harness-server/src/http.rs
@@ -18,6 +18,8 @@ use tokio::sync::{broadcast, RwLock};
 
 const MAX_WEBHOOK_BODY_BYTES: usize = 512 * 1024;
 
+mod task_routes;
+
 pub struct AppState {
     pub server: Arc<HarnessServer>,
     pub project_root: std::path::PathBuf,
@@ -94,13 +96,18 @@ pub async fn build_app_state(server: Arc<HarnessServer>) -> anyhow::Result<AppSt
         log_retention_days = server.config.observe.log_retention_days,
         "harness: effective config"
     );
-    if matches!(
-        server.config.server.github_webhook_secret.as_deref(),
-        Some("")
-    ) {
-        tracing::warn!(
-            "server.github_webhook_secret is configured as empty string; refusing webhook requests until this is set to a non-empty value"
-        );
+    match server.config.server.github_webhook_secret.as_deref() {
+        Some("") => {
+            tracing::warn!(
+                "server.github_webhook_secret is configured as empty string; refusing webhook requests until this is set to a non-empty value"
+            );
+        }
+        None => {
+            tracing::warn!(
+                "server.github_webhook_secret is not configured; refusing webhook requests until this is set to a non-empty value"
+            );
+        }
+        Some(_) => {}
     }
 
     let db_path = dir.join("tasks.db");
@@ -235,56 +242,6 @@ fn resolve_reviewer(
     (None, config.clone())
 }
 
-#[derive(Debug)]
-enum EnqueueTaskError {
-    BadRequest(String),
-    Internal(String),
-}
-
-async fn enqueue_task(
-    state: &Arc<AppState>,
-    req: task_runner::CreateTaskRequest,
-) -> Result<task_runner::TaskId, EnqueueTaskError> {
-    if req.prompt.is_none() && req.issue.is_none() && req.pr.is_none() {
-        return Err(EnqueueTaskError::BadRequest(
-            "at least one of prompt, issue, or pr must be provided".to_string(),
-        ));
-    }
-
-    let agent =
-        if let Some(name) = &req.agent {
-            state.server.agent_registry.get(name).ok_or_else(|| {
-                EnqueueTaskError::BadRequest(format!("agent '{name}' not registered"))
-            })?
-        } else {
-            state
-                .server
-                .agent_registry
-                .default_agent()
-                .ok_or_else(|| EnqueueTaskError::Internal("no agent registered".to_string()))?
-        };
-
-    let (reviewer, review_config) = resolve_reviewer(
-        &state.server.agent_registry,
-        &state.server.config.agents.review,
-        agent.name(),
-    );
-
-    let task_id = task_runner::spawn_task(
-        state.tasks.clone(),
-        agent,
-        reviewer,
-        review_config,
-        state.skills.clone(),
-        state.events.clone(),
-        state.interceptors.clone(),
-        req,
-    )
-    .await;
-
-    Ok(task_id)
-}
-
 pub async fn serve(server: Arc<HarnessServer>, addr: SocketAddr) -> anyhow::Result<()> {
     tracing::info!("harness: HTTP server listening on {addr}");
 
@@ -316,7 +273,7 @@ pub async fn serve(server: Arc<HarnessServer>, addr: SocketAddr) -> anyhow::Resu
         .route("/health", get(health_check))
         .route("/rpc", post(handle_rpc))
         .route("/ws", get(crate::websocket::ws_handler))
-        .route("/tasks", post(create_task))
+        .route("/tasks", post(task_routes::create_task))
         .route("/tasks", get(list_tasks))
         .route("/tasks/{id}", get(get_task))
         .route(
@@ -344,58 +301,43 @@ async fn handle_rpc(State(state): State<Arc<AppState>>, Json(req): Json<RpcReque
     }
 }
 
-async fn create_task(
-    State(state): State<Arc<AppState>>,
-    Json(req): Json<task_runner::CreateTaskRequest>,
-) -> (StatusCode, Json<serde_json::Value>) {
-    match enqueue_task(&state, req).await {
-        Ok(task_id) => (
-            StatusCode::ACCEPTED,
-            Json(json!({
-                "task_id": task_id.0,
-                "status": "running"
-            })),
-        ),
-        Err(EnqueueTaskError::BadRequest(error)) => {
-            (StatusCode::BAD_REQUEST, Json(json!({ "error": error })))
-        }
-        Err(EnqueueTaskError::Internal(error)) => (
-            StatusCode::INTERNAL_SERVER_ERROR,
-            Json(json!({ "error": error })),
-        ),
-    }
-}
-
 async fn github_webhook(
     State(state): State<Arc<AppState>>,
     headers: HeaderMap,
     body: Bytes,
 ) -> (StatusCode, Json<serde_json::Value>) {
-    if let Some(secret) = state.server.config.server.github_webhook_secret.as_deref() {
-        if secret.is_empty() {
+    let secret = match state.server.config.server.github_webhook_secret.as_deref() {
+        Some("") => {
             return (
                 StatusCode::INTERNAL_SERVER_ERROR,
                 Json(json!({"error": "invalid server.github_webhook_secret configuration"})),
-            );
+            )
         }
-        let signature = match headers
-            .get("x-hub-signature-256")
-            .and_then(|value| value.to_str().ok())
-        {
-            Some(signature) => signature,
-            None => {
-                return (
-                    StatusCode::UNAUTHORIZED,
-                    Json(json!({"error": "missing header x-hub-signature-256"})),
-                )
-            }
-        };
-        if !crate::webhook::verify_github_signature(secret, signature, body.as_ref()) {
+        Some(secret) => secret,
+        None => {
+            return (
+                StatusCode::INTERNAL_SERVER_ERROR,
+                Json(json!({"error": "missing server.github_webhook_secret configuration"})),
+            )
+        }
+    };
+    let signature = match headers
+        .get("x-hub-signature-256")
+        .and_then(|value| value.to_str().ok())
+    {
+        Some(signature) => signature,
+        None => {
             return (
                 StatusCode::UNAUTHORIZED,
-                Json(json!({"error": "invalid webhook signature"})),
-            );
+                Json(json!({"error": "missing header x-hub-signature-256"})),
+            )
         }
+    };
+    if !crate::webhook::verify_github_signature(secret, signature, body.as_ref()) {
+        return (
+            StatusCode::UNAUTHORIZED,
+            Json(json!({"error": "invalid webhook signature"})),
+        );
     }
 
     let event = match headers
@@ -437,7 +379,7 @@ async fn github_webhook(
         req.project = Some(state.project_root.clone());
     }
 
-    match enqueue_task(&state, req).await {
+    match task_routes::enqueue_task(&state, req).await {
         Ok(task_id) => (
             StatusCode::ACCEPTED,
             Json(json!({
@@ -446,10 +388,10 @@ async fn github_webhook(
                 "task_id": task_id.0,
             })),
         ),
-        Err(EnqueueTaskError::BadRequest(error)) => {
+        Err(task_routes::EnqueueTaskError::BadRequest(error)) => {
             (StatusCode::BAD_REQUEST, Json(json!({ "error": error })))
         }
-        Err(EnqueueTaskError::Internal(error)) => (
+        Err(task_routes::EnqueueTaskError::Internal(error)) => (
             StatusCode::INTERNAL_SERVER_ERROR,
             Json(json!({ "error": error })),
         ),
@@ -478,407 +420,4 @@ async fn get_task(State(state): State<Arc<AppState>>, Path(id): Path<String>) ->
 }
 
 #[cfg(test)]
-mod tests {
-    use super::*;
-    use async_trait::async_trait;
-    use axum::body::Body;
-    use axum::http::Request;
-    use harness_core::{
-        AgentRequest, AgentResponse, Capability, CodeAgent, StreamItem, TokenUsage,
-    };
-    use std::sync::Arc;
-    use tokio::sync::Mutex;
-    use tower::ServiceExt;
-
-    struct CapturingAgent {
-        prompts: Mutex<Vec<String>>,
-    }
-
-    impl CapturingAgent {
-        fn new() -> Arc<Self> {
-            Arc::new(Self {
-                prompts: Mutex::new(Vec::new()),
-            })
-        }
-    }
-
-    #[async_trait]
-    impl CodeAgent for CapturingAgent {
-        fn name(&self) -> &str {
-            "capturing-agent"
-        }
-
-        fn capabilities(&self) -> Vec<Capability> {
-            vec![]
-        }
-
-        async fn execute(&self, req: AgentRequest) -> harness_core::Result<AgentResponse> {
-            self.prompts.lock().await.push(req.prompt);
-            Ok(AgentResponse {
-                output: String::new(),
-                stderr: String::new(),
-                items: vec![],
-                token_usage: TokenUsage {
-                    input_tokens: 0,
-                    output_tokens: 0,
-                    total_tokens: 0,
-                    cost_usd: 0.0,
-                },
-                model: "mock".into(),
-                exit_code: Some(0),
-            })
-        }
-
-        async fn execute_stream(
-            &self,
-            _req: AgentRequest,
-            _tx: tokio::sync::mpsc::Sender<StreamItem>,
-        ) -> harness_core::Result<()> {
-            Ok(())
-        }
-    }
-
-    async fn make_test_state_with(
-        dir: &std::path::Path,
-        config: harness_core::HarnessConfig,
-        agent_registry: harness_agents::AgentRegistry,
-    ) -> anyhow::Result<Arc<AppState>> {
-        let thread_manager = crate::thread_manager::ThreadManager::new();
-        let server = Arc::new(crate::server::HarnessServer::new(
-            config,
-            thread_manager,
-            agent_registry,
-        ));
-        let tasks = task_runner::TaskStore::open(&dir.join("tasks.db")).await?;
-        let events = Arc::new(harness_observe::EventStore::new(dir)?);
-        let signal_detector = harness_gc::SignalDetector::new(
-            server.config.gc.signal_thresholds.clone().into(),
-            harness_core::ProjectId::new(),
-        );
-        let draft_store = harness_gc::DraftStore::new(dir)?;
-        let gc_agent = Arc::new(harness_gc::GcAgent::new(
-            harness_gc::gc_agent::GcConfig::default(),
-            signal_detector,
-            draft_store,
-        ));
-        let thread_db = crate::thread_db::ThreadDb::open(&dir.join("threads.db")).await?;
-        Ok(Arc::new(AppState {
-            server,
-            project_root: dir.to_path_buf(),
-            tasks,
-            skills: Arc::new(tokio::sync::RwLock::new(harness_skills::SkillStore::new())),
-            rules: Arc::new(tokio::sync::RwLock::new(
-                harness_rules::engine::RuleEngine::new(),
-            )),
-            events,
-            gc_agent,
-            plans: Arc::new(tokio::sync::RwLock::new(std::collections::HashMap::new())),
-            thread_db: Some(thread_db),
-            plan_db: None,
-            interceptors: vec![],
-            notification_tx: tokio::sync::broadcast::channel(32).0,
-            notification_lagged_total: Arc::new(AtomicU64::new(0)),
-            notification_lag_log_every: 1,
-            notify_tx: None,
-            initialized: Arc::new(AtomicBool::new(true)),
-        }))
-    }
-
-    async fn make_test_state(dir: &std::path::Path) -> anyhow::Result<Arc<AppState>> {
-        make_test_state_with(
-            dir,
-            harness_core::HarnessConfig::default(),
-            harness_agents::AgentRegistry::new("test"),
-        )
-        .await
-    }
-
-    async fn make_test_state_with_agent(
-        dir: &std::path::Path,
-        webhook_secret: Option<&str>,
-    ) -> anyhow::Result<(Arc<AppState>, Arc<CapturingAgent>)> {
-        let mut config = harness_core::HarnessConfig::default();
-        config.server.github_webhook_secret = webhook_secret.map(ToString::to_string);
-
-        let capturing = CapturingAgent::new();
-        let mut registry = harness_agents::AgentRegistry::new("test");
-        registry.register("test", capturing.clone());
-
-        let state = make_test_state_with(dir, config, registry).await?;
-        Ok((state, capturing))
-    }
-
-    fn webhook_app(state: Arc<AppState>) -> Router {
-        Router::new()
-            .route(
-                "/webhook",
-                post(github_webhook).layer(DefaultBodyLimit::max(MAX_WEBHOOK_BODY_BYTES)),
-            )
-            .with_state(state)
-    }
-
-    #[tokio::test]
-    async fn health_endpoint_returns_ok_and_task_count() -> anyhow::Result<()> {
-        let dir = tempfile::tempdir()?;
-        let state = make_test_state(dir.path()).await?;
-
-        let app = Router::new()
-            .route("/health", get(health_check))
-            .with_state(state);
-
-        let response = app
-            .oneshot(Request::builder().uri("/health").body(Body::empty())?)
-            .await?;
-
-        assert_eq!(response.status(), StatusCode::OK);
-
-        #[derive(serde::Deserialize, Debug)]
-        struct HealthResponse {
-            status: String,
-            tasks: u64,
-        }
-
-        use http_body_util::BodyExt;
-        let body = response.into_body().collect().await?.to_bytes();
-        let health: HealthResponse = serde_json::from_slice(&body)?;
-
-        assert_eq!(health.status, "ok");
-        assert_eq!(health.tasks, 0);
-        Ok(())
-    }
-
-    #[tokio::test]
-    async fn webhook_issue_mention_creates_issue_task() -> anyhow::Result<()> {
-        let dir = tempfile::tempdir()?;
-        let (state, _agent) = make_test_state_with_agent(dir.path(), None).await?;
-        let before_count = state.tasks.count();
-        let app = webhook_app(state.clone());
-
-        let payload = serde_json::json!({
-            "action": "created",
-            "issue": { "number": 106 },
-            "comment": { "body": "@harness please handle this issue" },
-            "repository": { "full_name": "majiayu000/harness" }
-        });
-
-        let response = app
-            .oneshot(
-                Request::builder()
-                    .method("POST")
-                    .uri("/webhook")
-                    .header("x-github-event", "issue_comment")
-                    .header("content-type", "application/json")
-                    .body(Body::from(payload.to_string()))?,
-            )
-            .await?;
-
-        assert_eq!(response.status(), StatusCode::ACCEPTED);
-        assert_eq!(state.tasks.count(), before_count + 1);
-        Ok(())
-    }
-
-    #[tokio::test]
-    async fn webhook_review_on_pr_creates_pr_review_task() -> anyhow::Result<()> {
-        let dir = tempfile::tempdir()?;
-        let (state, _agent) = make_test_state_with_agent(dir.path(), None).await?;
-        let before_count = state.tasks.count();
-        let app = webhook_app(state.clone());
-
-        let payload = serde_json::json!({
-            "action": "created",
-            "issue": { "number": 42, "pull_request": { "url": "https://api.github.com/repos/majiayu000/harness/pulls/42" } },
-            "comment": { "body": "@harness review" },
-            "repository": { "full_name": "majiayu000/harness" }
-        });
-
-        let response = app
-            .oneshot(
-                Request::builder()
-                    .method("POST")
-                    .uri("/webhook")
-                    .header("x-github-event", "issue_comment")
-                    .header("content-type", "application/json")
-                    .body(Body::from(payload.to_string()))?,
-            )
-            .await?;
-
-        assert_eq!(response.status(), StatusCode::ACCEPTED);
-        assert_eq!(state.tasks.count(), before_count + 1);
-        Ok(())
-    }
-
-    #[tokio::test]
-    async fn webhook_fix_ci_on_pr_creates_fix_ci_task() -> anyhow::Result<()> {
-        let dir = tempfile::tempdir()?;
-        let (state, _agent) = make_test_state_with_agent(dir.path(), None).await?;
-        let before_count = state.tasks.count();
-        let app = webhook_app(state.clone());
-
-        let payload = serde_json::json!({
-            "action": "created",
-            "issue": {
-                "number": 42,
-                "html_url": "https://github.com/majiayu000/harness/pull/42",
-                "pull_request": { "url": "https://api.github.com/repos/majiayu000/harness/pulls/42" }
-            },
-            "comment": {
-                "body": "@harness fix CI",
-                "html_url": "https://github.com/majiayu000/harness/issues/42#issuecomment-1"
-            },
-            "repository": { "full_name": "majiayu000/harness" }
-        });
-
-        let response = app
-            .oneshot(
-                Request::builder()
-                    .method("POST")
-                    .uri("/webhook")
-                    .header("x-github-event", "issue_comment")
-                    .header("content-type", "application/json")
-                    .body(Body::from(payload.to_string()))?,
-            )
-            .await?;
-
-        assert_eq!(response.status(), StatusCode::ACCEPTED);
-        assert_eq!(state.tasks.count(), before_count + 1);
-        Ok(())
-    }
-
-    #[tokio::test]
-    async fn webhook_secret_requires_signature_header() -> anyhow::Result<()> {
-        let dir = tempfile::tempdir()?;
-        let (state, _agent) = make_test_state_with_agent(dir.path(), Some("secret")).await?;
-        let app = webhook_app(state);
-
-        let payload = serde_json::json!({
-            "action": "created",
-            "issue": { "number": 106 },
-            "comment": { "body": "@harness please handle this issue" },
-            "repository": { "full_name": "majiayu000/harness" }
-        });
-
-        let response = app
-            .oneshot(
-                Request::builder()
-                    .method("POST")
-                    .uri("/webhook")
-                    .header("x-github-event", "issue_comment")
-                    .header("content-type", "application/json")
-                    .body(Body::from(payload.to_string()))?,
-            )
-            .await?;
-
-        assert_eq!(response.status(), StatusCode::UNAUTHORIZED);
-        Ok(())
-    }
-
-    #[tokio::test]
-    async fn webhook_secret_rejects_invalid_signature_value() -> anyhow::Result<()> {
-        let dir = tempfile::tempdir()?;
-        let (state, _agent) = make_test_state_with_agent(dir.path(), Some("secret")).await?;
-        let app = webhook_app(state);
-
-        let payload = serde_json::json!({
-            "action": "created",
-            "issue": { "number": 106 },
-            "comment": { "body": "@harness please handle this issue" },
-            "repository": { "full_name": "majiayu000/harness" }
-        });
-
-        let response = app
-            .oneshot(
-                Request::builder()
-                    .method("POST")
-                    .uri("/webhook")
-                    .header("x-github-event", "issue_comment")
-                    .header(
-                        "x-hub-signature-256",
-                        "sha256=0000000000000000000000000000000000000000000000000000000000000000",
-                    )
-                    .header("content-type", "application/json")
-                    .body(Body::from(payload.to_string()))?,
-            )
-            .await?;
-
-        assert_eq!(response.status(), StatusCode::UNAUTHORIZED);
-        Ok(())
-    }
-
-    #[tokio::test]
-    async fn webhook_empty_secret_configuration_fails_closed() -> anyhow::Result<()> {
-        let dir = tempfile::tempdir()?;
-        let (state, _agent) = make_test_state_with_agent(dir.path(), Some("")).await?;
-        let app = webhook_app(state);
-
-        let payload = serde_json::json!({
-            "action": "created",
-            "issue": { "number": 106 },
-            "comment": { "body": "@harness please handle this issue" },
-            "repository": { "full_name": "majiayu000/harness" }
-        });
-
-        let response = app
-            .oneshot(
-                Request::builder()
-                    .method("POST")
-                    .uri("/webhook")
-                    .header("x-github-event", "issue_comment")
-                    .header("content-type", "application/json")
-                    .body(Body::from(payload.to_string()))?,
-            )
-            .await?;
-
-        assert_eq!(response.status(), StatusCode::INTERNAL_SERVER_ERROR);
-        Ok(())
-    }
-
-    #[tokio::test]
-    async fn webhook_rejects_invalid_event_header() -> anyhow::Result<()> {
-        let dir = tempfile::tempdir()?;
-        let (state, _agent) = make_test_state_with_agent(dir.path(), None).await?;
-        let app = webhook_app(state);
-
-        let payload = serde_json::json!({
-            "action": "created",
-            "issue": { "number": 106 },
-            "comment": { "body": "@harness please handle this issue" },
-            "repository": { "full_name": "majiayu000/harness" }
-        });
-
-        let response = app
-            .oneshot(
-                Request::builder()
-                    .method("POST")
-                    .uri("/webhook")
-                    .header("x-github-event", "Issue-Comment")
-                    .header("content-type", "application/json")
-                    .body(Body::from(payload.to_string()))?,
-            )
-            .await?;
-
-        assert_eq!(response.status(), StatusCode::BAD_REQUEST);
-        Ok(())
-    }
-
-    #[tokio::test]
-    async fn webhook_body_limit_rejects_large_payload() -> anyhow::Result<()> {
-        let dir = tempfile::tempdir()?;
-        let (state, _agent) = make_test_state_with_agent(dir.path(), None).await?;
-        let app = webhook_app(state);
-
-        let oversized = vec![b'a'; MAX_WEBHOOK_BODY_BYTES + 1024];
-        let response = app
-            .oneshot(
-                Request::builder()
-                    .method("POST")
-                    .uri("/webhook")
-                    .header("x-github-event", "issue_comment")
-                    .header("content-type", "application/json")
-                    .body(Body::from(oversized))?,
-            )
-            .await?;
-
-        assert_eq!(response.status(), StatusCode::PAYLOAD_TOO_LARGE);
-        Ok(())
-    }
-}
+mod tests;
